Fix width issue, but now the border doesn't work
If the browser window is large enough for the SVG to hit max-height, the
border fits the width of the screen instead of the width of the SVG,
leaving big gray spaces on either side.

This is absolutely ridiculous.  Modern CSS is a total mess.  Even doing
the simplest things, like wanting to center something, or have something
scale and maintain aspect ratio--without using pixel sizes, which no one
should do--is practically impossible.  There's always some little thing
that breaks, no matter how many of the umpteen different methods you try.